jobLog

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/jpaCode/jpaDao/JpaXxlJobInfoDao.java
@@ -1,14 +1,14 @@
 package com.xxl.job.admin.jpaCode.jpaDao;
 
-import com.xxl.job.admin.core.model.XxlJobInfo;
-import com.xxl.job.admin.jpaCode.model.XxlJobGroupEntity;
+
 import com.xxl.job.admin.jpaCode.model.XxlJobInfoEntity;
-import org.apache.ibatis.annotations.Param;
+
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import javax.transaction.Transactional;
 
@@ -25,7 +25,7 @@ public interface JpaXxlJobInfoDao extends JpaRepository<XxlJobInfoEntity,Integer
     @Transactional
     @Modifying
     @Query("delete from XxlJobInfoEntity x where x.id=?1")
-    public int delete(@Param("id") int id);
+    public Long delete(@Param("id") int id);
 
     @Query("select x from XxlJobInfoEntity x where x.jobGroup = ?1 ")
     public List<XxlJobInfoEntity> getJobsByGroup(@Param("jobGroup") int jobGroup);
@@ -40,6 +40,5 @@ public interface JpaXxlJobInfoDao extends JpaRepository<XxlJobInfoEntity,Integer
     @Modifying
     @Query("update XxlJobInfoEntity x set x.triggerLastTime=:#{#xxlJobInfoEntity.triggerLastTime},x.triggerNextTime=:#{#xxlJobInfoEntity.triggerNextTime},x.triggerStatus=:#{#xxlJobInfoEntity.triggerStatus} where x.id=:#{#xxlJobInfoEntity.id}")
     public int scheduleUpdate(@Param("xxlJobInfoEntity")XxlJobInfoEntity xxlJobInfoEntity);
-
 
 }
